Small edits
NoChromo is not for production use, and is only made to bootstrap device, with size > quality
Note that because of this, things are either, missing, wrong, bad, or weird. Even more so, many packages do not contain shared libs to save storage, while others have no manuals, or are 100% static. NoChromo is a very fucky solution to making chroots from very little.